+++ RobotCode/lib/miam_utils_arduino/PID.h
@@ -17,33 +17,33 @@
                 /// \param[in] Kd Derivative gain.
                 /// \param[in] Ki Integral gain.
                 /// \param[in] maxIntegral Maximum value of Ki * integral_
-                PID(double const& Kp, double const& Kd, double const& Ki, double const& maxIntegral);
+                PID(float const& Kp, float const& Kd, float const& Ki, float const& maxIntegral);
 
                 /// \brief Compute PID output.
                 /// \param[in] error Value of current error (current - target).
                 /// \param[in] dt Time since last call, used to compute integral and derivative.
-                double computeValue(double const& error, double const& dt);
+                float computeValue(float const& error, float const& dt);
 
                 /// \brief Compute PID output.
                 /// \param[in] error Value of current error (current - target).
                 /// \param[in] errorDerivative Derivative of the error term.
                 /// \param[in] dt Time since last call, used to compute integral and derivative.
-                double computeValue(double const& error, double const& errorDerivative, double const& dt);
+                float computeValue(float const& error, float const& errorDerivative, float const& dt);
 
                 /// \brief Reset the integral to a specific value (default 0)
-                void resetIntegral(double const& value = 0.0);
+                void resetIntegral(float const& value = 0.0);
 
-                double getCorrection();
-                double getIntegral();
+                float getCorrection();
+                float getIntegral();
 
             private:
-                double Kp_; ///< Proportional gain.
-                double Kd_; ///< Derivative gain.
-                double Ki_; ///< Integral gain.
-                double maxIntegral_; ///< Maximum integral value.
-                double integral_; ///< Current integral value.
-                double previousError_; ///< Previous error, used to compute derivative.
-                double lastCorrection_; ///< Last correction value computed.
+                float Kp_; ///< Proportional gain.
+                float Kd_; ///< Derivative gain.
+                float Ki_; ///< Integral gain.
+                float maxIntegral_; ///< Maximum integral value.
+                float integral_; ///< Current integral value.
+                float previousError_; ///< Previous error, used to compute derivative.
+                float lastCorrection_; ///< Last correction value computed.
         };
     }
 #endif
